devKitARM v46 support, build system improvements
-Fixed devoptabs for devKitARM v46
-Discovered that the rules for the files that needed to be compiled in
ARM mode were not being fleshed out... seems automake doesn't like
make's ability to have multiple rules to expand variables and
dependencies. Fixed this mess by biting the bullet and switching to use
libtool to make a convenience library to tweak CFLAGS for it alone.
-Added a pkg-tool file.

+++ src/ctr_elf_loader.c
@@ -23,14 +23,14 @@ void ctr_load_header(Elf32_Ehdr *header, FILE *file)
 static int set_position(FILE *file, uint64_t position)
 {
 	if (fseek(file, 0, SEEK_SET)) return -1;
-	while (position > LONG_MAX)
+	while (position > INT_MAX)
 	{
-		long pos = LONG_MAX;
+		int pos = INT_MAX;
 		if (fseek(file, pos, SEEK_CUR)) return -1;
-		position -= LONG_MAX;
+		position -= INT_MAX;
 	}
 
-	if (fseek(file, position, SEEK_CUR)) return -1;
+	if (fseek(file, (int)position, SEEK_CUR)) return -1;
 	return 0;
 }
